Check for `cargo` earlier than for `make`
The problem is that nowadays `cargo` projects pretty often use `make`
internally for building native C libraries so `Makefile` is present in
the project root.

There is no 100% correct resolution on which method specific project uses:
- `make` uses `cargo`
- `cargo` uses `make`

But the latter variant seems to be used more often that's the reason for
change.